taxiAPI/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.8.2